chore: make default cfg
