+++ .gitignore
@@ -27,3 +27,4 @@ wire_gen.go
 
 __debug_bin
 .vscode
+coma.cfg
